fix(gql): notification map
